AwsCryptographyPrimitives/runtimes/python/pyproject.toml
[tool.poetry]
name = "aws-cryptography-primitives"
version = "0.1.0"
description = ""
authors = ["AWS Crypto Tools <aws-crypto-tools@amazon.com>"]
packages = [
    { include = "aws_cryptography_primitives", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.11.0"
# TODO: Depend on PyPi once Smithy-Python publishes their Python package there
smithy-python = { path = "../../../submodules/smithy-python/python-packages/smithy-python", develop = false}
DafnyRuntimePython = "^4.4.0"
standard-library = {path = "../../../StandardLibrary/runtimes/python", develop = false}
cryptography = "^41.0.3"
six = "^1.16.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
